:sparkles: JSON dependency injection
Make the `lua-cjson` module an optional dependency.
It's now possible to use other JSON libraries.
The library will try to `require` one of `cjson`, `dkjson`, `json`
automatically. It's also possible to inject the json library while
making the `QueryBuilder:execute()` call.
Replace `cjson` by `dkjson` for tests since it's a pure Lua module so we
don't need to compile it.